[ISSUE #129]🥚Define macros to handle common code for RequestHeader
